# docs/swagger.yaml
basePath: /
definitions:
...
  errors.ErrorValidation:
    properties:
      field:
        type: string
      reason:
        type: string
      reason_value:
        type: string
    type: object
  event.EventParticipantRequest:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  event.EventRequest:
    properties:
      description:
        type: string
      end_date:
        example: YYYY-MM-DDTHH:mm:ss.SSSZ
        type: string
      location:
        type: string
      name:
        type: string
      start_date:
        example: YYYY-MM-DDTHH:mm:ss.SSSZ
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - description
    - end_date
    - location
    - name
    - start_date
    type: object
  event.EventResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Event'
        type: array
      message:
        type: string
    type: object
  model.Event:
    properties:
      contact_person:
        $ref: '#/definitions/model.User'
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      start_date:
        type: string
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  model.Response:
    properties:
      message:
        type: string
    type: object
  model.ResponseWithError:
    properties:
      error:
        items:
          $ref: '#/definitions/errors.ErrorValidation'
        type: array
      message:
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  title: Event Management
  version: "1.0"
paths:
...
  /event:
    get:
      consumes:
      - application/json
      description: Find all Event with its data
      operationId: event-find-all
      parameters:
      - description: Filter with Tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Find All Event
      tags:
      - Event
    post:
      consumes:
      - application/json
      description: Create an Event
      operationId: event-create
      parameters:
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/event.EventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Create an Event
      tags:
      - Event
  /event/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an Event with their data
      operationId: event-delete
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Delete an Event
      tags:
      - Event
    patch:
      consumes:
      - application/json
      description: Update an Event with their data
      operationId: event-update
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/event.EventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Update an Event
      tags:
      - Event
  /event/{id}/attend/{code}:
    patch:
      consumes:
      - application/json
      description: Attend an Event
      operationId: event-attend
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Participant Code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Attend an Event
      tags:
      - Event
  /event/{id}/cancel/{code}:
    delete:
      consumes:
      - application/json
      description: Cancel participation of an Event
      operationId: event-cancel
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Participant Code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Cancel participation of an Event
      tags:
      - Event
  /event/{id}/participate:
    post:
      consumes:
      - application/json
      description: Create a participant with their data for an Event
      operationId: event-create-participant
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/event.EventParticipantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Create a participant for an Event
      tags:
      - Event
...
securityDefinitions:
  ApiBearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"